God. Please make this the last commit

+++ src/GameCoreInterface.java
@@ -131,7 +131,7 @@ public interface GameCoreInterface {
      * @param shopId The ID of the shop the player is selling an item to
      * @param item The item the player is selling (eventually will be an Item obj)
      */
-    public int sellItem(String name, int shopId, String item);
+    public double sellItem(String name, int shopId, String item);
     
     /**
      * 605B_buy_method
